# tyst-api-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tyst_api_rest",
    "description": "TYST REST API",
    "license": {
      "name": "Apache-2.0 WITH FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/se/{algorithm}/keygen": {
      "post": {
        "tags": [
          "se_resources"
        ],
        "summary": "Start key pair generation.",
        "description": "Start generation of a suitable asymmetric key pair for the signature algorithm.",
        "operationId": "se_keygen",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "description": "The signature algorithm.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "strength",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "strength": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Key generation has started. Redirecting to result resource."
          },
          "404": {
            "description": "Failed to start"
          },
          "429": {
            "description": "Too many requests are already in progress"
          }
        }
      }
    },
    "/se/{algorithm}/keygen/{generation_request_id}": {
      "get": {
        "tags": [
          "se_resources"
        ],
        "summary": "Get generated key pair.",
        "description": "Get generated suitable asymmetric key pair for the signature algorithm.",
        "operationId": "se_keygen_result",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "description": "The signature algorithm.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "generation_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "private_key",
                    "public_key_b64"
                  ],
                  "properties": {
                    "private_key": {
                      "type": "object",
                      "required": [
                        "key_material_b64"
                      ],
                      "properties": {
                        "confinement_id": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Identifier used by provider to locate the correct confinement."
                        },
                        "confinement_type": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Type of confinement. E.g. \"soft\"."
                        },
                        "key_material_b64": {
                          "type": "array",
                          "items": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 0
                          },
                          "description": "Base64 encoded key material"
                        },
                        "object_reference": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Identifier used by provider to locate the correct object."
                        }
                      }
                    },
                    "public_key_b64": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "307": {
            "description": "Generation is still in progress."
          },
          "404": {
            "description": "Key pair generation failed or was never started."
          }
        }
      }
    },
    "/se/{algorithm}/sign": {
      "post": {
        "tags": [
          "se_resources"
        ],
        "summary": "Generate a signature.",
        "description": "Generate a signature of the message (data) with the private key from the asymmetric key pair.",
        "operationId": "se_sign",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "private_key",
                  "message_b64"
                ],
                "properties": {
                  "message_b64": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    }
                  },
                  "private_key": {
                    "type": "object",
                    "required": [
                      "key_material_b64"
                    ],
                    "properties": {
                      "confinement_id": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Identifier used by provider to locate the correct confinement."
                      },
                      "confinement_type": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Type of confinement. E.g. \"soft\"."
                      },
                      "key_material_b64": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "format": "int32",
                          "minimum": 0
                        },
                        "description": "Base64 encoded key material"
                      },
                      "object_reference": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Identifier used by provider to locate the correct object."
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "signature_b64"
                  ],
                  "properties": {
                    "signature_b64": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failure"
          }
        }
      }
    },
    "/se/{algorithm}/verify": {
      "post": {
        "tags": [
          "se_resources"
        ],
        "summary": "Verify a signature.",
        "description": "Verify a signature of the message (data) with the public key from the asymmetric key pair.",
        "operationId": "se_verify",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "public_key_b64",
                  "signature_b64",
                  "message_b64"
                ],
                "properties": {
                  "message_b64": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    }
                  },
                  "public_key_b64": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    }
                  },
                  "signature_b64": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Signature verification failed"
          }
        }
      }
    },
...
  },
...
}